Radically refactor SchemaVisitor, visitSchema, and SchemaDirectiveVisitor.
Instead of a single SchemaDirectiveVisitor base class with a static
visitSchema method, there is now a much simpler abstract SchemaVisitor
base class that defines the visitor method signatures as well as a static
method called implementsVisitorMethod.

The logic for visiting schemas has been extracted into a new standalone
function called visitSchema. This visitSchema function accepts a
visitorSelector function that returns an array of visitors that are both
capable of and interested in visiting the current type.

The visitSchema function is used by the SchemaDirectiveVisitor class to
implement its own static visitSchemaDirectives method, which is almost
identical to the previous SchemaDirectiveVisitor.visitSchema method.

This refactoring should allow the core schema visiting logic to be reused
for other purposes.

In recognition of this more general purpose, the directives.ts module has
been renamed to schemaVisitor.ts.